# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Users and Warehouse API
  version: 1.0.0
paths:
  /warehouse:
    get:
      summary: Get all warehouse products
      tags:
        - warehouse
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WarehouseProduct'
    post:
      summary: Create a new warehouse product
      tags:
        - warehouse
      requestBody:
        description: Product to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarehouseProductRequest'
      responses:
        '201':
          description: Created product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseProduct'
        '400':
          description: Bad request
  /warehouse/{product_id}:
    get:
      summary: Get product by ID
      tags:
        - warehouse
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseProduct'
        '404':
          description: Product not found
    patch:
      summary: Update product parameters
      tags:
        - warehouse
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Product updates
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarehouseProductUpdate'
      responses:
        '200':
          description: Updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseProduct'
        '404':
          description: Product not found
        '400':
          description: Bad request
    delete:
      summary: Delete product
      tags:
        - warehouse
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Product deleted
        '404':
          description: Product not found
...
components:
  schemas:
    WarehouseProduct:
      type: object
      properties:
        product_id:
          type: string
          x-go-name: ProductID
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        quantity:
          type: integer
          minimum: 0
      required:
        - product_id
        - name
        - price
        - quantity
    WarehouseProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        quantity:
          type: integer
      required:
        - name
        - description
        - price
        - quantity
    WarehouseProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        quantity:
          type: integer
          minimum: 0
...